Add files via upload
The following are the files used to create the game. Quiz_Game.py is the main file. Game Questions.txt is exactly what you think it is, it holds the questions and answers to each question on the quiz. Sorry in advance if my code is messy or has inadequate commenting, my excuse is being lazy in the summer. setup.py uses cx_Freeze to build the Quiz_Game.py file into an executable.